+++ src/rpc/index.ts
@@ -1,3 +1,4 @@
 export * from "./adapter";
+export * from "./streaming-adapter";
 export * from "./router-builder";
 export * from "./utils";